Remove leaked master-entry relocation
