+++ src/main/java/uk/ac/ebi/embl/gff3tools/cli/FileConversionCommand.java
@@ -12,12 +12,10 @@ package uk.ac.ebi.embl.gff3tools.cli;
 
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.LoggerContext;
-import com.fasterxml.jackson.databind.MapperFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.json.JsonMapper;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
 import java.nio.file.AtomicMoveNotSupportedException;
@@ -33,21 +31,14 @@ import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import picocli.CommandLine;
-import uk.ac.ebi.embl.api.entry.Entry;
-import uk.ac.ebi.embl.flatfile.reader.ReaderOptions;
-import uk.ac.ebi.embl.flatfile.reader.embl.EmblEntryReader;
 import uk.ac.ebi.embl.gff3tools.Converter;
 import uk.ac.ebi.embl.gff3tools.exception.CLIException;
-import uk.ac.ebi.embl.gff3tools.exception.ExitException;
 import uk.ac.ebi.embl.gff3tools.exception.FormatSupportException;
 import uk.ac.ebi.embl.gff3tools.exception.NonExistingFile;
 import uk.ac.ebi.embl.gff3tools.exception.ReadException;
 import uk.ac.ebi.embl.gff3tools.fftogff3.FFToGff3Converter;
 import uk.ac.ebi.embl.gff3tools.fftogff3.FastaToGff3Converter;
 import uk.ac.ebi.embl.gff3tools.gff3toff.Gff3ToFFConverter;
-import uk.ac.ebi.embl.gff3tools.metadata.EmblEntryMetadataSource;
-import uk.ac.ebi.embl.gff3tools.metadata.MasterEntryJsonMetadataSource;
-import uk.ac.ebi.embl.gff3tools.metadata.MasterMetadata;
 import uk.ac.ebi.embl.gff3tools.metadata.MasterMetadataProvider;
 import uk.ac.ebi.embl.gff3tools.sequence.fasta.header.FastaHeaderProvider;
 import uk.ac.ebi.embl.gff3tools.tsvconverter.TSVToGFF3Converter;
@@ -275,63 +266,6 @@ public class FileConversionCommand extends AbstractCommand {
             return new FastaToGff3Converter(engine, inputFastaSource, minGapLength, gapType, linkageEvidence);
         } else {
             throw new FormatSupportException(fromFileType, toFileType);
-        }
-    }
-
-    private MasterMetadataProvider buildMetadataProvider(Path masterEntryPath) throws ExitException {
-        MasterMetadataProvider provider = new MasterMetadataProvider();
-        if (masterEntryPath == null) {
-            return provider;
-        }
-        String ext = getFileExtension(masterEntryPath).orElse("").toLowerCase();
-        switch (ext) {
-            case "json" -> provider.addSource(parseMasterEntryJson(masterEntryPath));
-            case "embl", "ff" -> provider.addSource(new EmblEntryMetadataSource(parseMasterEntryEmbl(masterEntryPath)));
-            default ->
-                throw new CLIException("Unrecognized --master-entry file extension '." + ext
-                        + "'. Supported: .json (MasterEntry JSON), .embl/.ff (EMBL flatfile).");
-        }
-        return provider;
-    }
-
-    private MasterEntryJsonMetadataSource parseMasterEntryJson(Path path) throws ExitException {
-        if (!Files.exists(path)) {
-            throw new NonExistingFile("The --master-entry file does not exist: " + path, null);
-        }
-        try {
-            ObjectMapper mapper = JsonMapper.builder()
-                    .enable(MapperFeature.ACCEPT_CASE_INSENSITIVE_PROPERTIES)
-                    .build();
-            MasterMetadata meta = mapper.readValue(path.toFile(), MasterMetadata.class);
-            return new MasterEntryJsonMetadataSource(meta);
-        } catch (NoSuchFileException e) {
-            throw new NonExistingFile("The --master-entry file does not exist: " + path, e);
-        } catch (IOException e) {
-            throw new ReadException(
-                    "Failed to read --master-entry JSON file '%s': %s".formatted(path, e.getMessage()), e);
-        }
-    }
-
-    private Entry parseMasterEntryEmbl(Path path) throws ExitException {
-        if (!Files.exists(path)) {
-            throw new NonExistingFile("The --master-entry file does not exist: " + path, null);
-        }
-        try (BufferedReader reader = Files.newBufferedReader(path)) {
-            ReaderOptions readerOptions = new ReaderOptions();
-            readerOptions.setIgnoreSequence(true);
-            EmblEntryReader entryReader =
-                    new EmblEntryReader(reader, EmblEntryReader.Format.EMBL_FORMAT, "master_reader", readerOptions);
-            Entry masterEntry = null;
-            while (entryReader.read() != null && entryReader.isEntry()) {
-                masterEntry = entryReader.getEntry();
-            }
-            if (masterEntry == null) {
-                throw new CLIException("No entry found in --master-entry EMBL file: " + path);
-            }
-            return masterEntry;
-        } catch (IOException e) {
-            throw new ReadException(
-                    "Failed to read --master-entry EMBL file '%s': %s".formatted(path, e.getMessage()), e);
         }
     }
